+++ .codesandbox/sandbox/src/end-to-end-test-helpers.ts
@@ -1,25 +1,7 @@
-import {
-  acp_v2 as acp,
-  createSolidDataset,
-  getSolidDataset,
-  saveSolidDatasetAt,
-  setThing,
-  getSourceUrl,
-  deleteSolidDataset,
-  createContainerAt,
-  deleteFile,
-} from "@inrupt/solid-client";
+import { createContainerAt, deleteFile } from "@inrupt/solid-client";
 import { Session } from "@inrupt/solid-client-authn-browser";
 
 export function getHelpers(podRoot: string, session: Session) {
-  const policyResourceUrl = getPolicyResourceUrl();
-
-  function getPolicyResourceUrl(baseUrl: string = podRoot) {
-    return (
-      baseUrl +
-      `solid-client-tests/browser/acp/policies-${session.info.sessionId}.ttl`
-    );
-  }
   function getTestContainerUrl(baseUrl: string = podRoot) {
     return (
       baseUrl +
@@ -37,112 +19,7 @@ export function getHelpers(podRoot: string, session: Session) {
     await deleteFile(getTestContainerUrl(), { fetch: session.fetch });
   }
 
-  async function initialisePolicyResource() {
-    let publicRule = acp.createRule(policyResourceUrl + "#rule-public");
-    publicRule = acp.setPublic(publicRule, true);
-
-    let publicReadPolicy = acp.createPolicy(
-      policyResourceUrl + "#policy-publicRead"
-    );
-    publicReadPolicy = acp.addRequiredRuleUrl(publicReadPolicy, publicRule);
-    publicReadPolicy = acp.setAllowModes(publicReadPolicy, {
-      read: true,
-      append: false,
-      write: false,
-    });
-
-    let selfRule = acp.createRule(policyResourceUrl + "#rule-self");
-    selfRule = acp.addAgent(selfRule, session.info.webId!);
-    let selfWriteNoReadPolicy = acp.createPolicy(
-      policyResourceUrl + "#policy-selfWriteNoRead"
-    );
-    selfWriteNoReadPolicy = acp.addRequiredRuleUrl(
-      selfWriteNoReadPolicy,
-      selfRule
-    );
-    selfWriteNoReadPolicy = acp.setAllowModes(selfWriteNoReadPolicy, {
-      read: false,
-      append: true,
-      write: true,
-    });
-    selfWriteNoReadPolicy = acp.setDenyModes(selfWriteNoReadPolicy, {
-      read: true,
-      append: false,
-      write: false,
-    });
-
-    let policyResource = createSolidDataset();
-    policyResource = setThing(policyResource, publicRule);
-    policyResource = setThing(policyResource, publicReadPolicy);
-    policyResource = setThing(policyResource, selfRule);
-    policyResource = setThing(policyResource, selfWriteNoReadPolicy);
-
-    return saveSolidDatasetAt(policyResourceUrl, policyResource, {
-      fetch: session.fetch,
-    });
-  }
-
-  async function fetchPolicyResourceUnauthenticated(baseUrl: string = podRoot) {
-    // Explicitly fetching this without passing the Session's fetcher,
-    // to verify whether public Read access works:
-    return getSolidDataset(getPolicyResourceUrl(baseUrl), {
-      fetch: window.fetch.bind(window),
-    });
-  }
-
-  async function fetchPolicyResourceAuthenticated() {
-    // Explicitly fetching this without passing the Session's fetcher,
-    // to verify whether public Read access works:
-    return getSolidDataset(getPolicyResourceUrl(), { fetch: session.fetch });
-  }
-
-  async function setPolicyResourcePublicRead() {
-    return applyPolicyToPolicyResource("policy-publicRead");
-  }
-
-  async function setPolicyResourceSelfWriteNoRead() {
-    return applyPolicyToPolicyResource("policy-selfWriteNoRead");
-  }
-
-  async function applyPolicyToPolicyResource(policyFragmentIdentifier: string) {
-    const resourceWithAcr = await acp.getSolidDatasetWithAcr(
-      policyResourceUrl,
-      {
-        fetch: session.fetch,
-      }
-    );
-    if (!acp.hasAccessibleAcr(resourceWithAcr)) {
-      throw new Error(
-        `The test Resource at [${getSourceUrl(
-          resourceWithAcr
-        )}] does not appear to have a readable Access Control Resource. Please check the Pod setup.`
-      );
-    }
-    const changedResourceWithAcr = acp.addPolicyUrl(
-      resourceWithAcr,
-      policyResourceUrl + "#" + policyFragmentIdentifier
-    );
-    return acp.saveAcrFor(changedResourceWithAcr, {
-      fetch: session.fetch,
-    });
-  }
-
-  async function deletePolicyResource() {
-    return deleteSolidDataset(policyResourceUrl, { fetch: session.fetch });
-  }
-
-  function getSessionInfo() {
-    return session.info;
-  }
-
   return {
-    initialisePolicyResource,
-    fetchPolicyResourceUnauthenticated,
-    fetchPolicyResourceAuthenticated,
-    setPolicyResourcePublicRead,
-    setPolicyResourceSelfWriteNoRead,
-    deletePolicyResource,
-    getSessionInfo,
     createContainer,
     deleteContainer,
   };
